# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Management System API
  version: 1.0.0
  description: |
    API documentation for the Medical Management System backend.

    This API provides endpoints for managing users, doctors, patients, appointments, and medical records.
    Authentication is required for most endpoints using JWT tokens.
  contact:
    name: Medical Management System
    email: support@medical-system.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.medical-system.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      operationId: getUserProfile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Authentication
      summary: Update current user profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    put:
      tags:
        - Authentication
      summary: Change user password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User management endpoints
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (admin only)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      summary: Update user (admin only)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - Users
      summary: Delete user (admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
...
  /doctors:
    get:
      tags:
        - Doctors
      summary: Get all doctors
      operationId: getAllDoctors
      security: []
      parameters:
        - name: specialization
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'
  /doctors/{id}:
    get:
      tags:
        - Doctors
      summary: Get doctor by ID
      operationId: getDoctorById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doctors/{id}/availability:
    get:
      tags:
        - Doctors
      summary: Get doctor availability
      operationId: getDoctorAvailability
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Doctor availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorAvailability'
  /doctors/profile:
    put:
      tags:
        - Doctors
      summary: Update doctor profile (doctor only)
      operationId: updateDoctorProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorProfileUpdate'
      responses:
        '200':
          description: Doctor profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
  /doctors/availability:
    post:
      tags:
        - Doctors
      summary: Set doctor availability (doctor only)
      operationId: setAvailability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAvailability'
      responses:
        '201':
          description: Availability set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorAvailability'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfileUpdate:
      type: object
      required:
        - first_name
        - last_name
      properties:
        first_name:
          type: string
          minLength: 2
          maxLength: 50
          example: John
        last_name:
          type: string
          minLength: 2
          maxLength: 50
          example: Doe
        phone:
          type: string
          example: '+1234567890'
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        gender:
          type: string
          enum:
            - male
            - female
            - other
          example: male
        address:
          type: string
          example: 123 Main St, City, State 12345
    ChangePassword:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          example: oldpassword123
        new_password:
          type: string
          minLength: 6
          example: newpassword123
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum:
            - admin
            - doctor
            - patient
          example: patient
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: '+1234567890'
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        gender:
          type: string
          enum:
            - male
            - female
            - other
          example: male
        address:
          type: string
          example: 123 Main St, City, State 12345
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        license_number:
          type: string
          example: MD123456
        specialization:
          type: string
          example: Cardiology
        experience_years:
          type: integer
          example: 10
        consultation_fee:
          type: number
          format: decimal
          example: 150
        education:
          type: string
          example: MD from Harvard Medical School
        bio:
          type: string
          example: Experienced cardiologist with 10 years of practice
        is_available:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
    DoctorProfileUpdate:
      type: object
      properties:
        license_number:
          type: string
          example: MD123456
        specialization:
          type: string
          example: Cardiology
        experience_years:
          type: integer
          example: 10
        consultation_fee:
          type: number
          format: decimal
          example: 150
        education:
          type: string
          example: MD from Harvard Medical School
        bio:
          type: string
          example: Experienced cardiologist with 10 years of practice
        is_available:
          type: boolean
          example: true
    DoctorAvailability:
      type: object
      properties:
        id:
          type: integer
          example: 1
        doctor_id:
          type: integer
          example: 1
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          example: 1
        start_time:
          type: string
          format: time
          example: '09:00'
        end_time:
          type: string
          format: time
          example: '17:00'
        is_available:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
    SetAvailability:
      type: object
      required:
        - day_of_week
        - start_time
        - end_time
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          example: 1
        start_time:
          type: string
          format: time
          example: '09:00'
        end_time:
          type: string
          format: time
          example: '17:00'
        is_available:
          type: boolean
          example: true
...
    Success:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
        success:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid input provided
        message:
          type: string
          example: The request could not be processed
        success:
          type: boolean
          example: false
...